manager 뉴스 페이지 수정
